% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_id.R
\name{add_id}
\alias{add_id}
\title{Enrich source data with an ID}
\usage{
add_id(data)
}
\arguments{
\item{data}{Source data (data.frame or data.table)}
}
\value{
data.table object with a column named .id
}
\description{
\code{add_id} just ensures data is a data.table and adds a columns named .id
}
\details{
\code{add_id} in called behing the scence by \code{\link{crazyfy}}
if needed to ensure a valid id is present in source data.
Still, we recommand to start from a dataset already having an id, be it generated
by a call to \code{add_id} or already pre-existing to ensure correct source data enrichment.
}
\examples{
library(stranger)
data(iris)
(iris.id <- add_id(iris))
}
