% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.anomalies.R
\name{as.anomalies}
\alias{as.anomalies}
\title{Create a class anomalies from a (pre-filterd) stranger/singular}
\usage{
as.anomalies(x, ...)
}
\arguments{
\item{x}{stranger/singular (pre-filered) object}
}
\value{
object of class anomalies -- see also \code{\link{get_anomalies}}.
}
\description{
One can use either dplyr filter verb or subseting records with base "[" to manually define anomalies
Then \code{as.anomalies} can be used to create anomalies object to be used for instance for visualisation purpose.
}
\details{
Note that it is expected from the user to filter the object. If not done, all records are defined as anomalies.
This manual approach can be useful with some metrics and litterature recommandation,
 for instance using a cutpoint of 2 for 'abod' method.
}
