% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_datasets_.R
\docType{data}
\name{BudgetUK}
\alias{BudgetUK}
\title{Budget Shares of British Households}
\format{A dataframe containing 1.519 rows and 10 columns.
\itemize{
\item wfood. budget share for food expenditure
\item wfuel. budget share for fuel expenditure
\item wcloth. budget share for clothing expenditure
\item walc. budget share for alcohol expenditure
\item wtrans. budget share for transport expenditure
\item wother. budget share for other good expenditure
\item totexp. total household expenditure (rounded to the nearest 10 UK pounds sterling)
\item income. total net household income (rounded to the nearest 10 UK pounds sterling)
\item age. age of household head
\item children. number of children
}}
\source{
Blundell, Richard, Alan Duncan and Krishna Pendakur (1998) “Semiparametric estimation and consumer demand”, Journal of Applied Econometrics, 13(5), 435–462.
}
\usage{
data(BudgetUK)
}
\description{
Budget Shares of British Households
}
\references{
Journal of Applied Econometrics data archive : \url{http://qed.econ.queensu.ca/jae/}
}
\keyword{datasets}
