% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{cor}
\alias{cor}
\alias{cor.stranger}
\alias{cor.default}
\title{Cor}
\usage{
cor(x, ...)

\method{cor}{stranger}(x, method = c("pearson", "kendall", "spearman"), ...)

\method{cor}{default}(x, ...)
}
\arguments{
\item{x}{a stranger  object}

\item{method}{see base cor function}
}
\description{
Correlation for stranger objects
}
