% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_stuff.R
\name{filter.stranger}
\alias{filter.stranger}
\alias{filter.singular}
\title{dplyr methods}
\usage{
\method{filter}{stranger}(.data, ...)

\method{filter}{singular}(.data, ...)
}
\description{
dplyr methods
}
