% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exploration.R
\name{explore}
\alias{explore}
\title{Data Exploration plots}
\usage{
explore(data, type = "histogram", keep = NULL, drop = NULL, ...)
}
\arguments{
\item{data}{is the data frame containing the observations. Each row represents an observation and each variable is stored in one column.}

\item{keep}{character vector: names of columns to keep (filter)}

\item{drop}{character vector: names of columns to drop (filter)}

\item{\dots}{is additional arguments to be passed to internal
functions. Currently only col.fill and bins for histograms.}
}
\description{
This is a simple function that produced plots that may help you to understand the shape of your dataset.
Today, 3 plots are implemented;(1) a small histogram for each features, (2) a correlation matrix and
(3) a 2-Dimension Hexbin Frequency.
}
\examples{
\dontrun{
explore(BudgetUK, type ="hist")
explore(BudgetUK, type ="cor",drop="children")
}

}
