% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info.singular}
\alias{get_info.singular}
\title{singularized metrics information}
\usage{
\method{get_info}{singular}(x, ...)
}
\value{
(invisible) list with two componenents: metrics and standardizations. First slot consists in metrics attributes (corresponding weird info), second slot is a vector containing the names of aggregated standardized derived metrics.
}
\description{
Retrieve some informaiton on the content of a singularize object.
Note: always print (cat)  information, store the content if you want to programmatically access it.
}
\examples{
\dontrun{
library(dplyr)
info <-  iris \%>\% select(-Species) \%>\% crazyfy() \%>\% stranger() \%>\% singularize() \%>\% get_info()
}
}
