% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info.stranger}
\alias{get_info.stranger}
\title{stranger object  information}
\usage{
\method{get_info}{stranger}(x, simplify = TRUE, ...)
}
\value{
matrix - metrics attributes (corresponding weird info), one row per weird.
}
\description{
Retrieve some information on the content of a stranger object.
}
\examples{
\dontrun{
library(dplyr)
info <-  iris \%>\% select(-Species) \%>\% crazyfy() \%>\% stranger() \%>\% get_info()
info
}
}
