% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_anomaly.R
\name{investigate_feature_importance}
\alias{investigate_feature_importance}
\title{importFrom ranger ranger}
\usage{
investigate_feature_importance(
  data,
  id,
  score,
  anomaly_id,
  num.trees = NULL,
  ...
)
}
\description{
importFrom ranger ranger
}
