% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singularize.R
\name{singularize}
\alias{singularize}
\title{Singularize: derive metrics after scores noramlisation and eventual stacking}
\usage{
singularize(
  strangerObject,
  methods = c("norm", "rank"),
  stacks = c("max", "avg", "min", "damavg", "pruavg"),
  prefix = "N_anom",
  ...
)
}
\arguments{
\item{strangerObject}{stranger object (before singularization)}

\item{methods}{normalisation method to be used (several possible)}

\item{stacks}{stacking methods to be used (several possible)}

\item{prefix}{prefix string to be used for new columns names}
}
\description{
Prepare from weirds normalized/stacked metric
}
