% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weirds.R
\name{weirds_list}
\alias{weirds_list}
\title{Provides the list of available methods}
\usage{
weirds_list(onlynames = FALSE)
}
\arguments{
\item{onlynames}{- logical: should be return some details or only the nickname
of methods to be used in \code{\link{strange}}, \code{\link{stranger}},
\code{\link{weird}} or \code{\link{lucky_odds}}.}
}
\description{
Provides the list of available methods
}
