% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\docType{methods}
\name{cor,stranger-method}
\alias{cor,stranger-method}
\title{Correlation for stranger objects}
\usage{
\S4method{cor}{stranger}(x, method = c("pearson", "kendall", "spearman"))
}
\description{
describeIn cor Correlation method for stranger objects.
}
