% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_stuff.R
\name{filter.stranger}
\alias{filter.stranger}
\alias{filter.singular}
\title{dplyr methods}
\usage{
\method{filter}{stranger}(.data, ...)

\method{filter}{singular}(.data, ...)
}
\arguments{
\item{.data}{object (class: sranger or singular)}

\item{\dots}{filter definition}
}
\description{
Methods to use with dplyr filter function
}
