% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singularize.R
\name{singularize}
\alias{singularize}
\alias{singular}
\title{Normalize anomalies metrics and (eventually) stack them}
\usage{
singularize(strangerObject, methods = c("norm", "rank"), stacks = c("max",
  "avg", "min", "damavg", "pruavg"), prefix = "N_anom", ...)
}
\arguments{
\item{strangerObject}{An object build with \code{\link{stranger}} or \code{strangest}}

\item{methods}{character vector of standardization methods: norm (normalize) and rank (ranking version). By default, both methods are used.}

\item{stacks}{character vectors of stacked versions. Note that some methods require more than 2 weirds metrics to be used}

\item{prefix}{default prefix used to name generated new metrics}

\item{\dots}{Additional parameters - Currently not used.}
}
\description{
\code{singularize} derives normalized/standardized versions of different weirds contained in a stranger object so that they are directly comparable and then propose various aggregated measures (stacking).
}
